server/blog/blogapi/views.py
from rest_framework import generics
from blogapi.models import Author
from blogapi.serializers import AuthorSerializer

class AuthorList(generics.ListCreateAPIView):
    queryset = Author.objects.all()
    serializer_class = AuthorSerializer
class AuthorDetail(generics.RetrieveUpdateDestroyAPIView):
    queryset = Author.objects.all()
    serializer_class = AuthorSerializer